possibilita informar a expressão via linha de comando e utilizar números com casa decimal

+++ expressao-matematica/solver.h
@@ -4,6 +4,6 @@
 #include "ed.h"
 
 // Recebe a expressão no formato postfix e retorna resultado
-int solve_expression(queue_t* postfix);
+double solve_expression(queue_t* postfix);
 
 #endif // EXPRESSION_SOLVER_H
